refactored highlight.py into utils.py; refactored sys test for win into func; simplified utils._enable_vt_mode()
