+++ tests/test_main.py
@@ -5,10 +5,10 @@ from subprocess import PIPE, run
 
 import pytest
 
-import devtools.highlight
+import devtools.utils
 from devtools import Debug, debug
 from devtools.ansi import strip_ansi
-from devtools.highlight import use_highlight
+from devtools.utils import use_highlight
 
 
 @pytest.mark.xfail(sys.platform == "win32", reason="yet unknown windows problem")
@@ -314,17 +314,17 @@ def test_use_highlight_manually_set(monkeypatch):
 @pytest.mark.skipif(sys.platform == 'win32', reason='windows os')
 def test_use_highlight_auto_not_win(monkeypatch):
     monkeypatch.delenv('TEST_DONT_USE_HIGHLIGHT', raising=False)
-    monkeypatch.setattr(devtools.highlight, 'isatty', lambda _=None: True)
+    monkeypatch.setattr(devtools.utils, 'isatty', lambda _=None: True)
     assert use_highlight() is True
 
 
 @pytest.mark.skipif(sys.platform != 'win32', reason='not windows os')
 def test_use_highlight_auto_win(monkeypatch):
     monkeypatch.delenv('TEST_DONT_USE_HIGHLIGHT', raising=False)
-    monkeypatch.setattr(devtools.highlight, 'isatty', lambda _=None: True)
+    monkeypatch.setattr(devtools.utils, 'isatty', lambda _=None: True)
 
-    monkeypatch.setattr(devtools.highlight, 'color_active', False)
+    monkeypatch.setattr(devtools.utils, 'color_active', False)
     assert use_highlight() is False
 
-    monkeypatch.setattr(devtools.highlight, 'color_active', True)
+    monkeypatch.setattr(devtools.utils, 'color_active', True)
     assert use_highlight() is True
